// webpage/src/utils/types/permissions.ts
export type Permission =
    "list-permissions"
  | "create-permission"
  | "delete-permission"
  | "list-roles"
  | "create-role"
  | "delete-role"
  | "update-user"
  | "delete-user"
  | "create-admin"
  | "create-category"
  | "update-category"
  | "delete-category"
  | "create-contribution"
  | "update-contribution"
  | "delete-contribution"
  | "rank-contribution";